% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms2.r
\name{rms2}
\alias{rms2}
\title{Reverse 2-sample MR analysis}
\description{
This analysis is a wrapper around MR and colocalisation functions for systematically searching OpenGWAS for mediating pathways from genotype to trait
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{igd_id}}{The OpenGWAS dataset to analyse as the outcome GWAS}

\item{\code{info}}{Meta data for igd_id}

\item{\code{gwashits}}{Data from of GWAS hits}

\item{\code{rsid_scan}}{List of results of PheWAS search - one item for each GWAS hit}

\item{\code{coloc_data}}{Data used for colocalisation analysis - list of lists (rsid x candidate trait)}

\item{\code{coloc_result}}{Results from colocalisation analysis - list of data frames, one data frame for each region analysed}

\item{\code{candidate_instruments}}{List of instrument datasets for each candidate}

\item{\code{candidate_instruments_outcome}}{Dataframe of instrument-outcome associations}

\item{\code{mr_scan}}{Results from MR analysis - list of data frames with one data frame for each region}

\item{\code{mv_dat}}{Data and results from mvmr}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rms2$new()}}
\item \href{#method-extract_gwashits}{\code{rms2$extract_gwashits()}}
\item \href{#method-scan_rsid}{\code{rms2$scan_rsid()}}
\item \href{#method-colocalise_rsid}{\code{rms2$colocalise_rsid()}}
\item \href{#method-plot_coloc}{\code{rms2$plot_coloc()}}
\item \href{#method-coloc_scan}{\code{rms2$coloc_scan()}}
\item \href{#method-mr}{\code{rms2$mr()}}
\item \href{#method-mvmr}{\code{rms2$mvmr()}}
\item \href{#method-mvmr_lasso}{\code{rms2$mvmr_lasso()}}
\item \href{#method-clone}{\code{rms2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise the rms2 class for a particular outcome GWAS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$new(igd_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{igd_id}}{The OpenGWAS ID to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_gwashits"></a>}}
\if{latex}{\out{\hypertarget{method-extract_gwashits}{}}}
\subsection{Method \code{extract_gwashits()}}{
Extract GWAS hits for igd_id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$extract_gwashits(
  igd_id = self$igd_id,
  pval = 5e-08,
  kb = 10000,
  r2 = 0.001,
  pop = "EUR"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{igd_id}}{Default=self$igd_id}

\item{\code{pval}}{Significance threshold for detecting GWAS hit default=5e-8}

\item{\code{kb}}{Clumping parameter default=10000}

\item{\code{r2}}{Clumping parameter default=0.001}

\item{\code{pop}}{1000 genome super population to use for clumping default='EUR'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan_rsid"></a>}}
\if{latex}{\out{\hypertarget{method-scan_rsid}{}}}
\subsection{Method \code{scan_rsid()}}{
Search OpenGWAS for associations with a particular GWAS hit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$scan_rsid(rsid, pval = 5e-08, idlist = NULL, check_population = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{RSID}

\item{\code{pval}}{Significance threshold for candidate traits default=5e-8}

\item{\code{idlist}}{Which traits to search. When NULL (default) will search all available traits}

\item{\code{check_population}}{Restrict candidate traits to the same population as the igd_id. Default=TRUE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-colocalise_rsid"></a>}}
\if{latex}{\out{\hypertarget{method-colocalise_rsid}{}}}
\subsection{Method \code{colocalise_rsid()}}{
Perform colocalisation between outcome trait and a candidate trait for a particular locus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$colocalise_rsid(rsid, trait_id, radius = 50000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{RSID}

\item{\code{trait_id}}{trait_id}

\item{\code{radius}}{Radius around which to perform analysis default=50000}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_coloc"></a>}}
\if{latex}{\out{\hypertarget{method-plot_coloc}{}}}
\subsection{Method \code{plot_coloc()}}{
Plot region for a colocalisation analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$plot_coloc(rsid, trait_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{rsid}

\item{\code{trait_id}}{trait_id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coloc_scan"></a>}}
\if{latex}{\out{\hypertarget{method-coloc_scan}{}}}
\subsection{Method \code{coloc_scan()}}{
Perform colocalisation systematically for all detected candidate traits for a given region
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$coloc_scan(rsid, radius = 50000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{rsid i.e. one of the outcome GWAS hits}

\item{\code{radius}}{Radius around which to perform analysis default=50000}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mr"></a>}}
\if{latex}{\out{\hypertarget{method-mr}{}}}
\subsection{Method \code{mr()}}{
Perform MR analysis of all candidate traits for a particular rsid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$mr(
  rsid,
  steiger_filtering = TRUE,
  exclude_rsid_region = TRUE,
  radius = 250000,
  mrmethod = "mr_ivw"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{GWAS hit from which candidate traits are obtained}

\item{\code{steiger_filtering}}{Whether to perform steiger filtering default=TRUE}

\item{\code{exclude_rsid_region}}{Whether to exclude the known GWAS hit default=TRUE}

\item{\code{radius}}{Excluded region radius default=250000}

\item{\code{mrmethod}}{MR method default="mr_ivw"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvmr"></a>}}
\if{latex}{\out{\hypertarget{method-mvmr}{}}}
\subsection{Method \code{mvmr()}}{
Perform multivariable MR for all or some of the candidate traits related to a particular region
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$mvmr(rsid, traitlist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{RSID from which candidate traits are obtained}

\item{\code{traitlist}}{List of traits to use. Default (NULL) is to use all, but important to scrutinise and manually exclude those which appear to be synonymous with the outcome}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvmr_lasso"></a>}}
\if{latex}{\out{\hypertarget{method-mvmr_lasso}{}}}
\subsection{Method \code{mvmr_lasso()}}{
This function will take the output from mvmr, perform feature selection from the analysed traits using MVMR-LASSO, and then re-estimate the MVMR associations with only the retained traits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$mvmr_lasso(rsid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsid}}{Region from which MVMR analysis has been performed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rms2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
